Refactoring: CATR is better than previous model with removed punctuation The avatar needs the models in the constructor now (less copy paste)
