+++ avatar_models/captioning/catr/bleu_score.txt
@@ -1,5 +1,5 @@
 Use CATR Model with GPU cuda:0
-100%|██████████| 5000/5000 [40:46<00:00,  2.04it/s]
-{'testlen': 49716, 'reflen': 48813, 'guess': [49716, 44716, 39716, 34716], 'correct': [32723, 15943, 6969, 3138]}
-ratio: 1.0184991703029722
-MS COCO {'Bleu_1': 0.6581985678654627, 'Bleu_2': 0.48443109906975035, 'Bleu_3': 0.3453210064753295, 'Bleu_4': 0.24700063139069237}
+100%|██████████| 5000/5000 [41:22<00:00,  2.01it/s]
+{'testlen': 49738, 'reflen': 48949, 'guess': [49738, 44738, 39738, 34738], 'correct': [36164, 19360, 9311, 4448]}
+ratio: 1.0161188175447708
+MS COCO {'Bleu_1': 0.7270899513450335, 'Bleu_2': 0.5609297382602266, 'Bleu_3': 0.41931051233215355, 'Bleu_4': 0.3117035904333875}
